requirements.txt
chromadb==1.0.12
groq==0.28.0
huggingface-hub==0.33.0
numpy==2.3.0
python-dotenv==1.1.0
sentence-transformers==4.1.0
streamlit==1.46.0
